requirements.txt
newsapi-python==0.2.6
django-environ==0.8.1
pymongo==3.12.1